requirements.txt
gradio
numpy
scikit-learn
tensorflow
torch
transformers
scikit-image
scipy
opencv-python
Pillow
